Optimized to only use 2 webpages

+++ student_manage/students/views.py
@@ -2,6 +2,9 @@ from django.shortcuts import render, redirect
 from .models import Student
 from .forms import StudentForm
 from .forms import FilterForm
+from django.urls import reverse
+from django.views.decorators.http import require_POST
+from django.http import HttpResponseNotAllowed
 
 def student_list(request):
     if request.method == 'GET':
@@ -42,24 +45,25 @@ def add_student(request):
     return render(request, 'students/add_student.html', {'form': form})
 
 def delete_student(request, student_id):
-    student = Student.objects.get(id=student_id)
+    if request.method != 'POST':
+        return HttpResponseNotAllowed(['POST'])
+
+    student = get_object_or_404(Student, id=student_id)
     student.delete()
     return redirect('/')
-
 
 from django.shortcuts import render, redirect, get_object_or_404
 from .models import Student
 from .forms import StudentForm
 
-def edit_student(request, student_id):
-    student = get_object_or_404(Student, id=student_id)
+def update_student(request, student_id):
     if request.method == 'POST':
-        form = StudentForm(request.POST, instance=student)
-        if form.is_valid():
-            form.save()
-            return redirect('/')
-    else:
-        form = StudentForm(instance=student)
-    return render(request, 'students/edit_student.html', {'form': form, 'student': student})
-
-
+        student = get_object_or_404(Student, id=student_id)
+        student.first_name = request.POST.get('full_name', '').split(' ')[0]
+        student.last_name = request.POST.get('full_name', '').split(' ')[1] if len(request.POST.get('full_name', '').split(' ')) > 1 else ''
+        student.course = request.POST.get('course', '')
+        student.gender = request.POST.get('gender', '')
+        student.age = request.POST.get('age', 0)
+        student.save()
+        return redirect(reverse('student_list'))
+    return redirect(reverse('student_list'))
